Better messages for doc/stars/starHTML.idx

+++ trunk/ptclassic/mk/subdomain.mk
@@ -84,11 +84,16 @@ depend:
 	done
 
 doc/stars/starHTML.idx: $(wildcard doc/stars/*.htm)
-	@echo "Updating $@"
-	rm -f $@
-	(cd doc/stars; \
-	domain=`echo $(dir $(ME)) | sed s=/$$==`; \
-	echo "set TYCHO $(PTOLEMY)/tycho; \
-		source $(PTOLEMY)/tycho/lib/idx/tychoMakeIndex.tcl; \
-		ptolemyStarHTMLIndex \
-		\"\{Ptolemy $$domain $(notdir $(ME)) stars\}\"" | itclsh)
+	if [ -w $(dir $@) ]; then \
+		@echo "Updating $@"
+		rm -f $@
+		(cd doc/stars; \
+		domain=`echo $(dir $(ME)) | sed s=/$$==`; \
+		echo "set TYCHO $(PTOLEMY)/tycho; \
+			source $(PTOLEMY)/tycho/lib/idx/tychoMakeIndex.tcl; \
+			ptolemyStarHTMLIndex \
+			\"\{Ptolemy $$domain $(notdir $(ME)) stars\}\"" | \
+		itclsh); \
+	else \
+		echo "$(dir $@) does not exist, skipping"; \
+	fi
